Display link to install specific version of CMake.
When the user asks for a specific version of CMake that is available in
the SDK, but not installed locally, detect the error that gets printed,
parse the requested CMake version, and then display a hyperlink to
install that specific version.

Do not display link if the requested version does not exist as a
remote package. That would be useless as the install would fail
to find the package anyways.

Do not display link if the requested version exists as a local package.
This protect against version matching errors inside gradle plugin side
from corrupting the behavior here.

Bug: 111690559
Test: Added unit tests for MissingCmakeErrorHandler methods.

Change-Id: Ice7e1ba7b20bc417e94ddcc2889b8126554c5e4c
Signed-off-by: Emre Kultursay <emrekultursay@google.com>