Add partial dirtying support for tooltips.
This CL adds a subregion dirtying for opaqueRepaint. This allows the tooltip
component to only cause repaints to JComponents under the old + new position
of the tooltip (and the range vertical line).

This CL also adds a caching behavior to ProfilerLayeredPane setBounds method
to prevent unnecessary invalidates (which the tooltip causes indirectly). This
is needed to ensure that the tooltip doesn't force a repaint of the entire
profiler panel.

Bug: 110106585
Test: Verified with swingp.
Change-Id: Ib38c6678268941bd4edb986eb34d3802cfd7c616