Clipping feature
CL extends the algorithm by adding support for clipped shapes.

3 new booleans have been added to ShapeModel constructor:
hasClippingMask, shouldBreakMaskChain and isLastShape.

hasClippingMask is true for the shape in question if it is used to mask the
shapes above it in sketch. Used for creating the <group> tag in xml to
separate the clipped groups and assign them the appropriate clipping
paths.

shouldBreakMaskChain is true for the shape in question if it is the
first shape, after a chain of shapes on which a mask is applied, that doesn't
have the specific mask applied on it. It breaks the mask chain. Used for
marking the end of a group and then adding the shape to the <vector> tag
instead of the <group> tag.

isLastShape is true if the shape in question is the last element in a
list of SketchLayers. Used to mark the end of a possible clipping group
but only after the shape is added to the <group> tag.

Bug: N/A
Test: DrawableGeneratorTest.java
Change-Id: I2e375e6ca27dcde0e3fe28ea86c3e5c7325e8c18