Have SampleDataListener consider pending move, delete operations
Currently, when SampleDataListener responds to virtual file changes
it uses the file's parent to determine the affected
SampleDataResourceRepository. This was meant to be a workaround for
when a file is deleted and the corresponding AndroidFacet is no longer
available.

This doesn't work when deleting the sample data directory
itself, however, since the parent directory might belong to a different
module (as is the case for Blaze projects). SampleDataListener's logic
for responding to file movements is incorrect as well, since
SampleDataListener#fileMoved only considers the new parent of
the file that was moved when determining whether or not the file was
relevant to the SampleDataResourceRepository. This means that the
repository will be invalidated when you move a new file into the sample
data directory, but not when you move a file out of the directory.

This CL resolves both issues by refactoring SampleDataListener so that
it checks file relevance for move and delete operations before the
operation completes. This is accomplished via the fileChangePending()
function, which caches the SampleDataResourceRepository to be
invalidated after a pending file change is completed, and
pendingFilechangeComplete(), which actually performs the invalidation.

Change-Id: I18eda430c38b62d1a285ba1c76af34a79249eabc
Fixes: 113103259
Test: updates to SampleDataResourceRepositoryTest