Fix vscroll behavior in network details
Now that we show a formatted editor on the overview
tab, we need the data viewer to support vertical scrolling
(since the editor is forced to fit in the top 40% of that
panel).

However, previously, vscroll behavior is currently disabled
on that component, since it wasn't needed for the request/
response tabs.

In this CL, we enable it by default, and disable it
from the request/response tabs.

We also expose a UI utils class which was previously
private, since it is generically useful and, well,
we are using it now!

Change-Id: I64cd2f728cf4faa8c8e7fea54199b285dd948346
Fixes: 113124016
Fixes: 112471687
Test: N/A, visual changes only