Fix and unignore flaky test in CpuProfilerStage
CpuProfilerStageTest#traceFilesGeneratedPerTrace was flaky, because it tests
the feature of generating trace temporary files. The problem is these files
are not deleted until all the tests run, so if we run the test multiple
times or run other tests in parallel we might end up with conflicting, or
repeated, file names, which would cause the test to fail.

This should fix the test flakiness, recently observed in PSQ.

Fixes: 113102299
Test: Fixing a unit test
Change-Id: Ibdc93da80ad8882149ee77a54d46bd1817753c62