Fix janky C++ project view
If you selected app->cpp->includes in project view and then
resync then the include folders opened and closed a bunch of
times while AS tried to reselect the same Node (post Sync).

The issue was that equality/hashing wasn't right in the *Node
classes related to C++ includes so the right Node was never
found.

This change fixes those equals/hashCode implementations and
also verifies that a bunch of other class's equals/hashCode
work as well.

Bug: NA
Test: Adjusted unittests

Change-Id: I6943b4f23b841e54d9e8be30dfaa56d2ba0a7148