Add database and model support for AllocationSamplingInfo
This CL leverages the existing data polling and storage mechanism
to cache and query AllocatinoSamplingInfo coming from perfd.

Each time the sampling mode is changed via MemoryProfilerStage, the
setting is cached via Intellij's PropertiesComponent mechanism. The
sampling mode is fetched via a custom Updater to ensure that it is
alway consistent with what's stored in the database. (e.g. Changing
the mode from the CpuProfiler will also update the mode in MemoryProfiler,
which will be done in a future CL)

Bug: 113348662
Test: MemoryStatsTableTest#testInsertAndGetData
MemoryDataPollerTest#testGetDataInRange
MemoryDataPollerTest#testGetDataOutOfRange
MemoryProfilerStage#testAllocationSamplingRateUpdates
MemoryProfilerStage#testAllocationSamplingRateCorrectlyInitialized
MemoryProfilerStage#testAllocationSamplingModePersistsAcrossStages

Change-Id: I02e5c77c7b475e4c1190e42a0330d78b28409de3