Filter material libraries the same as androidx.
We originally wanted to be able to use the old design library with
androidx, and the new material library with non androidx.
However that doesn't work because the material library has
dependencies on the other support libraries.

This CL changes the components shown on the palette to follow the
androidx settings for the material libraries.

Bug: 113291474
Test: Unit test updated.
Change-Id: I23aa40a9bee6c20f90dcf6ad4abe31d97c357b03