Fix slider position in Hue Slider Component.
The Knob of HueSliderComponent jumps from right side to
left side when using right arrow key to slide right.

The reason is 0f and 1f in Hue domain are same. Setting
value to 0f/1f may jump the knob to 1f/0f in some cases.

To avoid that we need some logic to avoid changing value
from 0f to 1f or from 1f to 0f. But there are some bias
when converting knob position (integer) to hue value (float).

Change the type of HueSliderComponent from Float to Int can
avoid the bias issue. Which also makes more sense when using
arrow key to adjust the Hue value since the range of Hue
field is Integer.

Bug: 112648431
Test: Updated and passed HueSliderComponentTest
      Updated and passed ColorAdjustPanelTest
Change-Id: I264fab90157acdd3405d570e52cdeca565173a3a