Symbols support
This CL adds the symbols support feature to the existing engine

The SketchFile now holds an SymbolsLibrary object that stores all the
SketchSymbolMasters from each page in a HashMap. When iterating through the artboard
layers, if the layer is an instance of a symbol, a SymbolModel is
retrieved from the library using the SymbolId. Afterwards, the symbol is
converted into models and then generated just like a SketchShapeGroup

Symbol opacity, rotation and mirroring are not supported yet.

Bug: N/A
Test: DrawableGeneratorTest.java
Change-Id: Ie1ed6db251d02bf69424d5107e2870e2462a7806