Add new performance tests for sync
These new tests use the large and uber like projects.

Also, made some changes to the way the benchmarks are grouped, now
the measurements for the same project will be on the same benchmark,
making it easier to visualize.

With this change, there are 3 benchmarks:
 - Average
 - Large
 - Uber Like

And every benchmark has these metrics:
 - Initial IDEA
 - Initial Single Variant
 - Regular IDEA
 - Regular Single Variant

Test: New tests.
Bug: 113078294
Change-Id: I39295fd2ed923b8a5dba47aface2e755b8d8e8fd