Adopt the explicit package name from ExternalLibrary
Previously, we allowed the build system to attach a "representative
manifest" to External Library, which would be a valid manifest not
associated with the libary but containing the correct package name.

It is much simpler for the build system to just determine the package
name and attach it explicitly in such situations, so this CL replaces
the representativeManifest attribute with an optional explicit package
name.

Bug: 112667464
Test: Existing
Change-Id: I58043d09ad2b84ddc746e71a2ac01dd9ff4e2f38