Transformations support for symbols
This CL contains more features on symbols, and the restructuring and
bug-fixing that came along with them.

0. Added missing javadoc to classes and methods
1. Added a TransparentModel interface to centralize the properties and
methods of the models that can have custom transparency.
2. Added an InheritedProperties class to hold all the properties that
can be inherited from the parent groups, such as opacity, transformation
properties or constraints. Along with it, some transformation bugs have
been fixed.
3. Added support for symbols rotation and flipping, more accurate
scaling and rendering of symbols that contain other symbols inside.
4. Started implementing the ResizingConstraint class to support
constraint functionality. This class is not complete however and will
need more work to cover all the constraint cases and treat them
accordingly.

Symbols still have some transformation bugs yet to be fixed

Test: DrawableFileGeneratorTest.java
Bug: 113375629
Change-Id: Iacdb089287428f9d4f3d545c3f2be0f5d2af259f