Fix thread query perf with split query and in-mem cache
Log:
This CL is the result of a number of different approaches to optimizing the
"find last event prior to window" problem, which is encoded in the JOIN part
of the query, in the first half of the UNION ALL..

Modification #1:
* Add filtering to the outer table of the join.
* Add indexes and force FROM to use them (the query planner isn't very
  good).
* 350ms -> 250ms

Modification #2:
* Flip the outer and inner queries in the join.
* Remains at 350ms.

Modification #3:
* Complete the MAX query before doing the JOIN.
* Remains at 350ms.

Modification #4:
* Realize that SQLite is different from most other SQL implementations, as
it is not a host/client setup over a connection, but directly in the app.
In other words, it's better to rely on the host langauge/app/RT to perform
multiple queries than to do JOINs.
Therefore, the JOIN was removed and the query is split into two: first to
get all the distinct threadIds given the session and timestamp(1), and then
effectively perform a simple lookup of the MAX time's thread activity entry
for each threadId before the window, and UNION ALL with the thread
activities within the window.
(1) It turns out SQLite isn't great with having extra information either, so
the timestamp was omitted from the first query.
* 350ms -> 40ms.

Modification #5:
* The SELECT DISTINCT/GROUP BY for distinct threadIds is slow in SQLite,
even though an index is readily available. Apparently SQLite (or perhaps the
limitations of the SQL langauge spec) doesn't have an optimization when all
target rows fit within the index.
* A separate in-memory cache is set up for all unique threadIds (which calls
the first query in Modification #4 and populates the DB once if the cache is
not populated), and is used in place of the first query.
* 40ms -> 0.8ms.

Change-Id: I71374789551816a7aca904a448d47d17fb6f7155
Test: Expanded tests in prior CL and ran perf test to visualize speedup
Fixes: 110106452