Make Assistant content loading async
Loading the bundle from web or bundle from config can
take a long time, which caused UI to freeze, without an
indicator that it is loading either. Added the indicator
and moved content loading to a task to be dispatched,
with callback to display the contents after loading is done.

Bug: 113942257
Test: WhatsNewAssistantSidePanelTest
Change-Id: I87653c654939a494ee49dbdc4311df17c5dd08a3