Notify user when an unsupported version of Sketch was used
When importing a Sketch file, the user now gets an error notification
if the file is not valid or if it was created with an older version of
Sketch (anything before 50.0). For version 43.0 and higher, they can see
what version was used to create the file (extracted from meta.json).

Bug: N/A
Test: MetaParserTest.kt, manual for the UI
Change-Id: I3a1d1d2d0aa0841ecf316aedffee36855bc70251