Rewrite ParcelQuickFixTest, QuickFixMultiFileTest
This is a major rewrite which gets some tests passing with Bazel,
and significantly reduces their dependency on Kotlin test
framework code.

Note that AbstractQuickFixMultiFileTest and KotlinLightProjectDescriptor
are copied and adapted from the Kotlin test framework.

Unfortunately, about 13 of the ParcelQuickFix tests still fail, but
now they are comparison failures. These will be fixed in a later CL.

Test: Added
Bug: 114479438
Change-Id: I29913fa0ffbfb89fdb9975f8407625dd9550bd9a