Add logic for finding matching AndroidX dependencies.
When adding an artifact to a module, we want to select a version
that isn't known to cause version incompabilities with the
existing artifacts.

With androidX this is now a version range.
Added code to deal with ranges in general.

Always attempt to add the latest known version,
then go backwards in the list of known versions until a version
without incompabilities is found.

As an optimization stop when the added artifact is 2 or more
major versions older than an existing artifact found somewhere
else.

In this CL also convert a couple of recently added
AndroidGradleTests to more efficient AndroidTests.

Bug: 111656600
Test: Added unit tests.
Change-Id: I0a134bd5c9c672fe5d08d317e0a01205841a7771