Implement test R classes
- They are backed by the repository from
ResourceRepositoryManager#computeTestAppResources() implemented in
the previous CL, which for now only includes resources from the same
module under androidTest/res (this will be fixed soon).
- The R classes use the module's package name with the ".test" suffix,
as this is how Gradle generates them.
- They are only added to the resolve scope of androidTests, as determined
by TestArtifactSearchScopes which we use to distinguish local unit tests
from on-device instrumented tests.
- Implementations of main and test R classes share common code but
provide a different "resources source" which is a strategy object that
can be used by both the top-level class as well as inner (per resource
type) classes.
- After 3.3 I will refactor the AAR R classes (also based on
repositories) to use this mechanism, but this change is too big to
be cherry-picked easily.
- Renamed some related types, most noticeably *PackageRClass
and *ResourceTypeClass have been renamed to *RClass and *InnerRClass.

Bug: 116692965
Test: New AndroidGradleTestCase in LightClassesTests
Change-Id: If5494b30bad4e84792e700a00deb51b8746ff991