Update SceneManager render API to use Futures
The fix b/114307584 caused at least one of the LinearLayout tests to
fail. The reason is that, after this change, the modelChanged listener
would trigger before the model contained the information from Layoutlib.
This CL updates the SceneManager interface to allow setModel to wait on
the Layoutlib result before calling the modelChanged listeners.

Bug: 114307584
Test: Existing tests pass
Change-Id: Ied8d7fd9d845ce1ad0227aa617798fccc5df95d7