Fix assertion failure in IdeaSourceProvider
IdeaSourceProvider contains asserts that the renderscript and aidl
directories are not null. However, these methods are expected to
return null for projects that don't have such directories.

This patch replaces the assertion with a empty list return value,
which correctly indicates that there is no such directory for
the given project.

Test: Existing
Change-Id: I631a6f40aafe36bbcaaddbd022f3e240cd6d7bb0