Fix issue where the tooltip isn't updating
This CL fixes an issue where the tooltip isn't updating when the mouse
transitions from one monitor to another. This is caused by the change
that the tooltip no longer solely relies on the updater, and needs to
be attached to the highlight range manually. Now the tooltip explicitly
requires a dependent Range (or none if it's a fixed legend). As a
consequence, Updatable has been removed from LegendComponentModel.

Furthermore, value caching has been moved into LegendComponent, as
paint optimization is really a characteristic of an UI component, not
as part of the model.

This CL also fixes an issue with ResizingAxisComponent since it
it omitted caching the global range min in its global range min check.
This was causing unnecessary repaints that caused a performance
degradation with the new changes.

Change-Id: I035dd2fc551c967505201567db2f8db28d72751b
Fixes: 116771705
Test: Visually confirmed via swingp and passes updated tests
(cherry picked from commit 94a4150eb2b017ff9224c07632b6b1f828c9d658)