[AVD Mgr] Use the correct quick-fix for a missing Emulator
If the AVD Manager sees that the Emulator is missing, it says
"No emulator installed" and presents a clickable "Install Emulator"
link.

The link was bound to update/install platform tools; now it's
bound to update/install the Emulator.

I manually verified the correct behavior by removing the
<SDK>/emulator/ directory, verifying that the link appears,
clicking the link, and verifying that the Emulator is
downloaded.

Fixes: 116650173
Test: None
Change-Id: Ic304db41e085e424a33c6640569826c67555cd3e
(cherry picked from commit 05063f26045641d003d3591d5a1bfcb1dfa20001)