Fix AndroidResourceFoldingTest
This slightly modifies the change at go/aog/672555 so that
the app resource repository is created (if necessary) when
ResourceFoldingBuilder requests it.

If we don't want to compute app resources for ResourceFoldingBuilder in
general, then the alternative fix would be to compute them ahead of time
in the unit test.

Test: //tools/adt/idea/android-kotlin:intellij.android.kotlin.tests_tests
Bug: 114479438
Change-Id: I7b20293d64f42a9d1bd5f3f9af7b2ac143c42bcb