Fix race condition in SyncUtilTest
In waitForSmartAndSyncedAndCheckThread, there is a race between the
thread that schedules the runWhenSmartAndSync and the code checking for
the event to be in the EDT queue.
This fixes the problem by waiting the thread to start.

Change-Id: I67bd3a66f241db13706661b253105be67a74cbb8
Fixes: 117647243
Test: Updated test
(cherry picked from commit d7e46f8a47bbb50894a9b14b3787e0abef1cf172)