Filter out private view classes from tag names.
This CL will eliminate non public View classes as tag names for:
- DefaultXmlTagNameProvider
- LegacyCompletionContributor
the first used used at the start of a tag and the latter when
the user asks for completions in the middle of a tag name.

Our AndroidLayoutXmlTagNameProvider was already filtering out
these view classes, but we do not hide anything created by
the DefaultXmlTagNameProvider.
Our provider will still control the order and strikeout fonts
used for specific classes of tag names.

Fixes: 111843761
Test: Added a unit test

Change-Id: If30b51ec853755a37931192dc9be14be482772a3