Remove testAdaptive…Contents from TestGroup.UNRELIABLE
This is sort of an experiment. The test is failing frequently
in the continuous build, but we can't reproduce it on our workstations.
It happens to be running first in the TestGroup, which sometimes
stresses tests out or something. Back in TestGroup.DEFAULT it probably
won't run first, and maybe it will just pass. If not, oh well.
Other tests are failing at the moment.

Issue: 77269384
Change-Id: Ia47e641192346d08d35b4d7ff49708c16b5fad67