Update "AndroidDataBindingTest" with parameterized JUnit4 approach
Previously, this test used JUnit3, ultimately inheriting TestCase
through AndroidTestBase. AndroidTestBase, unfortunately, supports
many, many subclasses, which has caused it to grow with a fair
amount of awkward hardcoding, requiring overriding base methods
in fragile ways.

This new version is improved by two JUnit4 tricks:

1) Parameterization
Before, we had to create two subclasses, which were responsible for
setting the DataBindingMode before handling the rest of the work
back up to the parent class. Now, we can delete those classes.

2) AndroidProjectRule
Instantiating your own, isolated project rule is much cleaner, as
it allows you to know exactly how it's being used, without having
to jump back and forth between parent and child classes.

Finally - it turns out a base class in the JUnit3 version did some
tricks in order to make sure the tests ran on the EDT thread. Now,
we just handle that explicitly, which is clearer anyway as it
allows the reader to know which part of our tests need to run
from the event dispatch thread and which don't.

Bug: N/A
Test: Only test code updated

Change-Id: Ic0bf5e202c2d5810c5f8ffb4bf090d9de111ef09