Look up grpc client from stream id if the mapping exists in ProfilerService
In the new data pipeline, the GetCurrentTime and GetVersion should also
use the stream id (analogous to device id) to look up the corresponding
grpc client to redirect the call to.

Also updated the TimeRequest and VersionRequest to use the new stream_id
proto field (replacing the old device_id field).

Bug: 73538507
Test: Refactored and added UnifiedPipelineProfilerServiceTest
Change-Id: Ia67dc3fd2595fe468d847480252e6a52598313d7