[Gradle Sync] Prevent the parsing of build files from crashing sync.
This change ensures that certain methods in the DSL parser don't throw
any exception checked or unchecked and instead log the crash to
intellijs logs and the StudioCrashReporter but don't terminate the
thread. This has been added to prevent bad Gradle files from crashing
sync and rendering the IDE useless.

The Gradle build file parser is a hotspot for exceptions due to the
complexity of the DSL and how we attempt to mirror Gradles behaviour in
running these scripts.

These crashes will be logged and fixed but we won't render Android
Studio useless while doing so.

Bug: N/A
Test: N/A
Change-Id: Ib3f00aaf3fc92e7b3503005824ded7c8880602a7