Ensure sample data index consistency within layouts
This creates an index for sample data per parsed layout, which is
incremented every time a layout is used. The parser then uses that index
to select a given element for sample data if none were specified. This
ensures that the same element of sample data will be used in the same
place every time, and that there will be a correct increment of 1 when
using a LinearLayout or a RecyclerView.

Bug: 80051843
Test: tests added in LayoutPsiPullParserTest
Change-Id: I78fb5b4c8697f0fc24065663f1c4274cf51d3d6d