Ensure cleanup after cancelled SDK installation.
When an install operation is performed, it typically means unzipping
into the target location under Android SDK tree, which is not an atomic
operation. If it gets cancelled during unzipping, the tree is left in
an inconsistent state, which may confuse certain package locator helpers
and in general such a situation is better to be avoided.

Bug: b/68788981
Test: FullInstaller doesn't currently have a test class, the main
logic of AbstractPackageOperation is covered by
BasicInstallerTest.testCleanupWhenCancelled which is added in the
second CL in this topic (ag/5396077).
Change-Id: Ib074a878f4333062271e0c3d73a999218b5c0897
(cherry picked from commit 549547e472ae3388af89fc5f94987abc6c00316f)