IDEA-196754 2018.2 Android sync fails but UI does not show the failed status
The change provide the AndroidSyncIssue reporting with the sync task id to be able to work with the Build Sync view.
However it's only a workaround. It will not work for the parallel imports when multiple gradle projects are linked in the same IDE project.
The "real" fix would be to add the sync task id to the GradleSyncMessages api.

(cherry picked from commit 5db8be4d1dc5744a028f4f1105687697dac3975f)(IDEA-CR-39033)