Symbolize functions of native APKs built outside of Studio
Users have the option to profile APKs in Android Studio even if they were
not built using our IDE. For native apps, profiling such APKs using
simpleperf would result in useless callstacks, containing only entries in
the format "libsomething.so+0x001fd3". The reason for that is the symbols
shipped with APK are usually stripped to make it smaller. However, users
can provide the directories containing the unstripped symbol files when
profiling/debugging native APKs.

In this CL, we get the directories provided by the user and pass the one
corresponding to the architecture of the profiled process to simpleperf
report-sample command using the --symdir flag.

Bug: 117866211
Test: Added a couple of tests to SimpleperfSampleReporterTest:
      * unknownSymbolsSymbolizedWhenProvidingSymDir
      * unknownSymbolsNotSymbolizedWhenSymDirNotProvided
      * unknownSymbolsNotSymbolizedWhenProvidingInvalidSymDir
      The change can also be manually tested using the scenarios
      described in go/native-symbolizer-test

Change-Id: I02a6491417a892c26826975c8500afecc23092e1