Refresh schema on activate surface
When we activate the NavDesignSurface we do
NavigationSchema.quickValidate to see if there have been changes to the
known Navigators. However, this check won't find any newly-added
Navigators.
Now, in addition to the existing check, we always rebuild the schema in
the background when activating the design surface. The schema is only
actually updated if there are any changes.
In order for us to be able to update the ui in response to schema
changes, it is not possible to register listeners on NavigationSchema
that will be called when a schema is updated.
In this CL the listener mechanism is adopted in AddDestinationMenu.

Bug: 118826893. This doesn't completely fix the bug, since there are
some cases where we may need to refresh that aren't covered here.
However, I believe it covers the cases necessary for public release.
Test: Fixed problems with existing NavigationSchemaTest cases that were
exposed by this change. Added NavigationSchemaTest.testRebuildNoChange
and testListeners. Add NavDesignSurfaceTest.testActivateWithSchemaChange
and testActivateAddNavigator.

Change-Id: Ifc5a08ad4d552ac2a953cbf50c686127e49dda13