[AVD Mgr] Allow phones with bigger screens
We separate "phones" and "tablets" based on diagonal screen size.
We used to say that anything with a screen >= 6 inches is a tablet,
but the Pixel 3 XL phone's screen is 6.3 inches.

The smallest tablet we emulate is the 2012 Nexus 7, which is
7.0. Now anything < 7 is a phone and >= 7 is a tablet.

Also when creating a new AVD, make the Pixel 2 the default phone
(was Nexus 5x) and make the Pixel C the default tablet (was Nexus 9).

Patch set 2: Clean up compiler warnings.

Bug: 117556079
Test: Added ChooseSystemImagePanelTest.testPhoneVsTablet
Test: Updated DeviceMenuActionTest.testActions
Change-Id: I49b7252dc8008cb0e0b1a279a0f878a2b7c34902