Load custom view palette items in the background
During initialization, NlPaletteModel queries the PSI system for all
classes in the project that inherit from android.view.View. This
happens on the EDT and invloves file IO, and so it can lead to UI
freezes (as bad as 15 seconds for the Photos app in ASWB).

This CL resolves the issue by offloading the work for finding custom
view classes to a background thread and having NlPaletteModel trigger
a UI update asynchronously once the model has been updated to include
project components. Since the PSI query requires the global RW lock,
we use a non-blocking read action with retries to give priority to
incoming write actions which would otherwise freeze the UI.

NlPaletteModelTest has been migrated to use AndroidProjectRule instead
of extending AndroidTestCase so that the test logic no longer runs on
the EDT (otherwise having the test wait until the model signals an
update would lead to deadlock).

Change-Id: I2bea271da94ec44f91fc87b7def4a3ddca8c339e
Fixes: 118760540
Test: NlPaletteModelTest