Adding a latch to block stop until run completes.
A flaky test was found due to the test stopping before the poller completed. This
happened about 1% of the time locally. To get a 100% repo simply add a sleep to the
top of ProfilerService::getEventGroups.
This blocking of the stop function resolves the flake by ensuring the run function has
completed before continuing to tearing down the test.

Test: Fixed
Change-Id: Ib3695877718bd2d78d569ca133ba642cff4246b5
Fixes: 118904668
Fixes: 119043227