Storing a cache of custom queries instead of creating new ones.
This does two things. The first is a minor performance gain. The
second is prevents us from managing the lifetime of queries created.
The existing code would create a statement and needs the caller to
manage the lifetime. Now we reuse statements allowing us to open
it once and close it when the service terminates.

Test: UnifiedEventsTableTest verifies statement is same object.
Bug: 119134037
Change-Id: Ie1f2ea289f4b3edbfd9d32780da735e507fd47f5