Fix WNA always auto-showing without internet
When I resolved the merge conflict while rebasing
ag/5416195 I forgot that the change to parseBundle()
in ag/5424490 would also affect it. With the change
to not unpack the default resource, it meant that
only checking whether the local cache file exists no
longer works. It thought that since there was no cached
file that the "new" file was always a new version,
causing the WNA to auto-show every time without the
file from web.

Bug: 117521401
Test: WhatsNewAssistantBundleCreatorTest.testNewConfigVersion
Change-Id: Ie4194c2b1f2bdeeda7c47d0880b0b8fab9ca4a2b