[AVD Mgr] Don't look for Emulator in <sdk>/tools/
We now put the Emulator in <sdk>/emulator/ but in olden times,
we put it in <sdk>/tools/. The AVD Manager would look for
the Emulator in emulator/ first, and then in tools/. By now,
though, any Emulator image in tools/ is too old to be usable.

Don't look in tools/ for the Emulator.

Bug: 119327162
Test: Added AvdManagerConnectionTest.testFindEmulator
Change-Id: I710c88ddc141289cdff11b7e64a0eb14722c64d0