Split CreateCppKotlinProjectTest, handle flaky clipboard
CreateCppKotlinProjectTest timed out too much, so part of its flow was
split into a new BuildCppKotlinTest. The original
CreateCppKotlinProjectTest created AND built a test, and this was too
large to fit comfortably within the timeouts of a test.

The new CreateCppKotlinProjectTest now simply creates a project and
compares if the created project's files (mostly) match with a
pre-created project.  The pre-created project is composed of the files
we expect a new project to contain after being created.

The BuildCppKotlinTest imports the pre-created project and checks if the
project builds successfully.

Additionally, this commit contains a workaround around a bug regarding
the clipboard. The root bug is mentioned in the documentation for
com.intellij.Patches.SLOW_GETTING_CLIPBOARD_CONTENTS, which references
the JDK bug https://bugs.openjdk.java.net/browse/JDK-4818143. The
workaround implemented in IntelliJ was to create an XWinClipboardHandler
class inside the ClipboardSynchronizer class. The XWinClipboardHandler
class contains a "checkContentsQuick" method that checks if the system
clipboard contains data that is available to be read. If so, the method
returns a non-null value; if not, the method returns null. In case the
method returns null, the end result for a paste action is to do nothing.
This can cause a text field element to not change the contents of its
text field.

However, for a UI test, even though Jetbrains' workaround avoids a
deeper bug that appears in the JDK or even the X server, it can trip up
a test. A human user might just try pasting again, but the test does not
know if it should try again. The workaround for UI tests is to check if
the contents in the system clipboard match the value that was last
stored into the clipboard.

Fixes: 119283528
Fixes: 119132317
Test: bazel test //tools/adt/idea/android-uitests:CreateCppKotlinProjectTest
Change-Id: I10cd4530bb3ae25ee6ce5266eec626ec1f99478d