Track project size from background thread
The syncEnded callback happens on the EDT in Gradle projects. We don't
necessarily need to run from the EDT for this particular task, so we
just push it down to a background task.

Apart from moving to a background thread, we also use DumbService to
ensure that the computable is invoked when indices are available, and
handle possible exceptions that could arise.

Bug: 118603070
Test: manually tested that stats are reported
Change-Id: Ic6dce2941ad3ece91b27c96d8b3509ef3d608dac