Fix issue with whitespaces when running report-sample
When running simpleperf report-sample on the host, we had a problem with
whitespaces on the path because we're passing the command as a single string
to Runtime.getRuntime().exec() and it tries to infer the executable path by
checking where the first space is.

If your Android Studio path is "/mydir/studio/", the command to be run will
be something like "/mydir/studio/...../simpleperf report-sample .... ...."
and that will run just fine.

However, if your path is "/mydir/Android Studio/", the command will be
"/mydir/Android Studio/..../simpleperf report-sample .... ....". As we're
passing it as a single string to Runtime.getRuntime().exec, it will try to
stop at the first whitespace and consequently run "/mydir/Android", which is
definitely not what we want.

Using ProcessBuilder and a list of args fixes the issue, because we make the
args explicit and they don't need to be infered by checking when there is a
whitespace.

Fixes: 119540757
Bug: 119493028
Test: Modified affected test.

Change-Id: I76d85cf55f6b009eb10accdaa369ee8441e94849