Avoid modifying Palette in use.
Custom components are now loaded asynchronously. After a class is
found the cached palette used to be updated the newly found
components. That causes problem if the UI thread is iterating over
the existing components at the same time.

To avoid the above problem, this CL will create a new palette
before replacing the older cached version.

Change-Id: I103b25443ecec058ccb1c115139dae886d8fb8dc
Fixes: 119342694
Test: The test: PalettePanelTest#testSearchPaletteWithCustomComponent
      should stop failing sporadically.