Detect custom permissions by searching manifest PsiFile
When you perform a scoped search operation over Java classes,
AndroidResolveScopeEnlarger checks Android manifests for the presence of
custom permissions and, if they are present, ensures that the classes
generated for these permissions are included in the search scope. This
check relies on IntelliJ's XML DOM API, which has heavy initialization
costs (including creating resource repositories in some cases). This
leads to UI freezes when the EDT is blocked on the search operation.

This CL resolves the issue by manually searching the PsiFile
corresponding to the manifest for custom permission names instead of
relying on the more expensive XML DOM API.

Fixes: 119137052
Test: LightClassesTests

Change-Id: I92ac8c69b01ded297e8675799a6450599e0395c4