Execute retry, follow-up logic outside non-blocking read action
NlPaletteModel#loadAdditionalComponents uses a non-blocking read action
to query the PSI for custom views. Currently, both the follow-up logic
which processes the results of the read action and the logic that covers
retrying the read action if it's preempted are executed from inside the
read action itself, which means that we're holding on to the read lock
for longer than we need to. In the case of the follow-up logic, this
results in an unhandled ProcessCancelledException if a write action
comes along. In the case of the retry logic, this results in an
IllegalStateException because
ProgressIndicatorUtils#yieldToPendingWriteAction must not be called from
within a read action.

This CL resolves these issues by:

1. Making sure all the custom view processing logic happens inside the
computation passed to NonBlockReadAction#submit so that we can safely
call replaceProjectComponents outside the read action. We accomplish
this by introducing the CustomViewInfo data class, which stores the
results of the read action.

2. Spawning a separate background thread in the onSuccess callback
where we execute replaceProjectComponents (it would be nice if we could
just stay on the same thread and give up the read lock, but AFAICT there
isn't a simple way to do this).

3. Invoking the recursive retry of loadAdditionalComponents on the EDT.
This guarantees that the write action which preempted us has completed
by the time we retry the read action, but without attempting to yield
to write actions from inside the read action itself.

Change-Id: I38d44377fa069b0f10cb23418438b624c6a58ab0
Fixes: 119672436
Test: NlPaletteModelTest still passes