Add AndroidModuleSystem#getResourceModuleDependencies
Currently, ProjectResourceRepository decides which modules to create
ModuleResourceRepository instances for by using
AndroidUtils#getAllAndroidDependencies, which determines the module's
transitive Android dependencies based on the order entries in the
project's IML files. Depending on how the project's build system models
the underlying IntelliJ project, this may result in creating more
ModuleResourceRepository instances than necessary, contributing to UI
freezes during Layout Editor initialization.

This CL helps resolve the issue by introducing the
getResourceModuleDependencies() API to the module system abstraction,
allowing us to define build system-specific strategies for determining
the Android modules that a module transitively depends on for resources.
To help prevent erroneous calls to getAllAndroidDependencies() in the
context of resource resolution, this CL adds the
getAndroidResourceDependencies() utility to AndroidUtils. This method
delegates to the new module system API, and includes documentation about
when to use it over getAllAndroidDependencies(). ProjectResourceRepository
has been updated to use this new utility for determining which modules
to create ModuleResourceRepository instances for.

Gradle and default projects are accurately represented by the IML files,
so the implementations of getResourceModuleDependencies() in
GradleModuleSystem and DefaultModuleSystem just redirect to
AndroidUtils#getAllAndroidDependencies.

Bug: 119786187
Test: Redirects to existing method

Change-Id: I3afea1700a4e80221d09e4f288b4745b16ee929b