Remove Sessions flag and legacy logic for device+process selection
This change cleans up some legacy logic that supports the old, non-session
workflow. In particular, the separate device and process selection dropdowns
are removed. Also, we no longer need to handle the different device
scenarios inside NullMonitorStageView, as the info of the device is fully
embedded without the process selection dropdown.

Bug: 77649021
Test: Updated
Change-Id: I23c8e19576675dd978c66f2a4efa4099e652088b