Convert data binding tests to use EdtRule
Apparently the EdtRule is an alternate approach that lets us
avoid needing to use EdtTestUtils. As a disadvantage, we lose
the granularity of knowing which parts of the test need to run
on the EDT vs. which don't, but we do gain a bit of readability
by removing an indent.

Bug: N/A
Test: Updated
Change-Id: I9c5cab7d4257c5ca16e0c1906da33d6be28498c4