Add intention action to add maven.google.com dependencies
For various well-known classes (such as RecyclerView, WorkManager, etc),
if the corresponding maven.google.com Gradle artifact isn't already
one of the dependencies, offer to add it.

Longer term, we'll probably switch to a more generalized backend/index
of available classes, but this is a stop-gap to help discoverability
of dependencies in the meantime.

Still missing: Support for unresolved symbols in XML files.

Test: Integration tests included
Bug: N/A
Change-Id: I8fb41c1bc868c5f5aa78ded2aa5d33eb01cd4a75