Fix table update problem.
When editing a new item in the declared table, and then deleting the
row causes an IndexOfOuRange exception in the table update logic.

To fix this: remove the logic to restore the editor when no next
editor line is specified. At this point the model has already changed
and there is no way to restore the old editor.

Bug: 120441518
Test: Unit test added.

Change-Id: I6c771c79db6c51dc0cdcc329823ad8f63e3dbcc7