Fix for exporting API initiated traces.
This fix is a bit of a work around for the existing pipeline. The
proper fix for this would be to have the save function call out to
the datastore and load the capture into memory then save it to disk.
That change coupled with having an event (in the new pipeline) to know
when a capture is complete will allow for us to remove the need to save
a copy of the trace from the datastore.

Test: Updated CpuDataPoller to check for file path in trace info.
Change-Id: Idb21811d62c62a07c8b409f5b2084b3cee89eea5
Fixes: 120264801
(cherry picked from commit ba95039de75a8b20e720cdc9bea65d0e0be58632)