Don't cache Palette as side-effect of loading it
In I103b25443, we updated NlPaletteModel to load and modify a new
Palette in order to register third-party components and then replace the
existing Palette in the cache once the new Palette is ready. This was to
prevent the ConcurrentModificationExceptions we'd get when updating
the cached Palette asynchronously while it was still in use. The logic
for loading the new Palette still caches the Palette as a side-effect,
however, and so the exceptions never went away.

This CL resolves the issue by removing the caching side-effect in
NlPaletteModel#loadPalette. We are already explicitly calling
saveInPaletteMap() where caching is wanted, so no other changes to
NlPaletteModel are necessary. PreviewProviderTest and NlPaletteModelTest
have been updated so that they no longer rely on loadPalette for
caching.

Change-Id: Ifadbb4122510dbcbbb7fb3188367dbecfe3f5c0e
Fixes: 120445657
Fixes: 119342694
Test: PalettePanelTest + PreviewProviderTest + NlPaletteModelTest