[Gradle Files] Ensure commenting code triggers Sync Needed banner
To do this we refine the check for un-needed psi changes, override the before* methods on the listener to ensure we catch every useful change since some changes (such as adding a /) will sometimes not being detected.

We also perform some minor cleanup to remove possible NPEs and assertion failures.

Test: unit test added
Change-Id: I1de9b5cd5f06486027cd5b584fa1bec837e8f82d
Fixes: 119612082