Add empty TextRanges for our light classes and fields
During refactoring (move declaration, etc.) these classes will also
participate in finding usages, so we need to be able to build
UsageInfos out of them. By default, LightElement returns a null text
range, and that breaks the assumptions in UsageInfo constructors: a
text range is needed if (and only if) the containing file is defined
and it is not binary.

With this CL we make sure to provide an empty text range whenever we
override getContainingFile(); this means no navigation is available in
the preview tool window, but the refactoring does succeed.

Bug: 121057068
Bug: 128851036
Test: added
Change-Id: I6dd3612cd76a600003aca8182d8056ce6e4b1fae