Add Converter for app:layout_behavior
The app:layout_behavior defined by the material library is handled as
just a string but it is a reference to a class. This CL fixes the
attribute definition to use a class reference to a
CoordinatorLayout.Behavior so it is resolved correctly.
This both provides autocompletion and fixes the AndroidX refactoring
since this will now be recognized as a class reference.

Test: Updated AndroidX tests
Change-Id: I0f54d8cdf6b4144efe035c69dde0c94661d03dfa
Fixes: 116766487