Notify ResourceChangeListeners on file renamed
The ResourceNotificationManager was only dispatching event when a psi
child was modified.
When renaming a file, only the propertyChange method gets triggered and
the ResourceChangeListener was not notifing its listeners.

We were actually lucky that it sometimes worked before because if the file
is referenced from another file, a child*() method was triggered anyway,
for example:

class MainActivity {
   setLayoutContent(R.layout.abc)
}

When refactoring the name of "layout/abc.xml", the psi element
representing R.layout.abc was changed, hence triggering the childReplace()
of the PsiTree listeners with the parent being MainActivity.java

Test: ProjectResourcesBrowserViewModelTest#updateOnFileNameChanged
Test: ResourceNotificationManagerTest#testNotifyOnRename
Change-Id: If34409378a5f514a64feab56625b6e28f808165e
Fixes: 120933211