Avoid committing after action change.
The timing around focus loss seem to vary esspecially on Windows.
We were setting a temporary value and with the expected timing
the action should have overridden the temporary value.
This did not happen on Windows. So a better solution is to avoid
committing on focus loss after a an action.

Change-Id: I2b55d9ff72759d710d89682a015770a8118d777a
Fixes: 121350826
Test: Changed the unit test.