Update setting from user resizes only.
In tool windows we used to store the tool window widths every time
the WorkBench received a component resize event. Unfortunately
this event happens during initial display of tool windows on some
platforms.
In this CL make sure we only store the width of the tool windows
if the user made the adjustment.

Change-Id: I282fe8ef1ee6391897cfe8b53517204e5ffa4e1c
Fixes: 121258170
Test: Unit tests updated.