Make WNA resource config not required
A default config was needed in WNA resources as
a fallback when loading from web didn't work. Prior
to the load from web change, the fallback was to
open DAC release notes in browser instead, when the
local resource was not available. I changed this when
adding async load, under the assumption that there
would always be a local config to fall back on.

However, new releases such as early canaries don't
really need WNA notes, and we don't want to block
releases, so I'm removing this restriction. Now,
having a default resource is a prerequisite before
WNA even attempts to check the web for a config, so
we don't open a blank assistant panel or keep the
user waiting for the async loading without feedback.

Bug: N/A
Test: Existing, and removed 1 test
Change-Id: I2d170e70eb4b4e29fcbff150f1d2fd21e5d41076