Clean up DebugOnEmulatorTest
The SimpleLocalApplication test project was removed, but
DebugOnEmulatorTest still had a dependency on the test project. This
caused the test to fail when it tried to import the test project.

Also, the test requires more time on the CI machines because it copies
the Android SDK and starts the Android emulator. These two operations
can consume a large amount of time on the weaker CI machines. The size
of the test was set to enormous. This test continues to run only on the
QA team's machine.

Bug: 120862107
Test: bazel test //tools/adt/idea/android-uitests:DebugOnEmulatorTest
Change-Id: Ie125fe9d9325cee81c5335d02814dbd05d9fdcda