Guard against project disposal in DependencyManager.
In change: I1d1f1d we move the initialization onto a background
thread. The user may decide to close the project immediately
which would cause the crash from b/122373349 since the project
may not be valid when the message bus listener is setup.
Guard aganst that.

Also add a check in MigrateToAndroidxUtil.getProjectProperties
since this typically is called from a read transaction.
If the method is called from a background thread the project
may be disposed before the read transaction is started.

Cherry-pick from ag/5975384 with 1 line added to cleanup the
DependencyManager created in the newly added test.

Change-Id: I8bbe3d5260e421be7bcf46baf96372a9ad198227
Fixes: 122373349
Fixes: 122560560
Test: unit test added.