Proactively suggest new PSD when build files are being edited manually.
This CL introduces new editor notification to suggest
the new PSD as the alternative to configure project. Whenever
a build file is opened for editing, the editor notification
at the top will prompt the user to open Project Structure
instead.

This is in line with our strategy to reduce the role of manual
actions with the DSL in the IDE. We do the same for Translation
Editor, which is the preferred way to edit the string resources
(as opposed to direct XML manipulations).

Test: Extended ProjectSyncStatusNotificationProviderTest correspondingly.
Bug: b/122353862
Change-Id: I72a6eb9a3b1e11954d412e80c263b6f1850b3ab1